+++ src/test/java/com/alibaba/json/bvt/feature/FeaturesTest.java
@@ -1,6 +1,4 @@
-package com.alibaba.json.bvt;
-
-import java.util.List;
+package com.alibaba.json.bvt.feature;
 
 import org.junit.Assert;
 import junit.framework.TestCase;
@@ -10,14 +8,14 @@ import com.alibaba.fastjson.annotation.JSONField;
 import com.alibaba.fastjson.serializer.SerializeConfig;
 import com.alibaba.fastjson.serializer.SerializerFeature;
 
-public class FeaturesTest6 extends TestCase {
+public class FeaturesTest extends TestCase {
 
     public void test_0() throws Exception {
         SerializeConfig config = new SerializeConfig();
         config.setAsmEnable(false);
         
         String text = JSON.toJSONString(new Entity(), config);
-        Assert.assertEquals("{\"value\":[]}", text);
+        Assert.assertEquals("{\"value\":null}", text);
     }
     
     public void test_1() throws Exception {
@@ -25,18 +23,21 @@ public class FeaturesTest6 extends TestCase {
         config.setAsmEnable(true);
         
         String text = JSON.toJSONString(new Entity(), config);
-        Assert.assertEquals("{\"value\":[]}", text);
+        Assert.assertEquals("{\"value\":null}", text);
     }
 
     public static class Entity {
 
-        private List value;
+        private Object value;
 
-        @JSONField(serialzeFeatures = { SerializerFeature.WriteMapNullValue, SerializerFeature.WriteNullListAsEmpty })
-        public List getValue() {
+        @JSONField(serialzeFeatures = { SerializerFeature.WriteMapNullValue })
+        public Object getValue() {
             return value;
         }
 
+        public void setValue(Object value) {
+            this.value = value;
+        }
 
     }
 }
