+++ src/test/java/com/alibaba/json/bvt/feature/FeaturesTest3.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt;
+package com.alibaba.json.bvt.feature;
 
 import org.junit.Assert;
 import junit.framework.TestCase;
@@ -8,14 +8,14 @@ import com.alibaba.fastjson.annotation.JSONField;
 import com.alibaba.fastjson.serializer.SerializeConfig;
 import com.alibaba.fastjson.serializer.SerializerFeature;
 
-public class FeaturesTest7 extends TestCase {
+public class FeaturesTest3 extends TestCase {
 
     public void test_0() throws Exception {
         SerializeConfig config = new SerializeConfig();
         config.setAsmEnable(false);
         
         String text = JSON.toJSONString(new Entity(), config);
-        Assert.assertEquals("{\"value\":\"SECONDS\"}", text);
+        Assert.assertEquals("{\"value\":0}", text);
     }
     
     public void test_1() throws Exception {
@@ -23,22 +23,18 @@ public class FeaturesTest7 extends TestCase {
         config.setAsmEnable(true);
         
         String text = JSON.toJSONString(new Entity(), config);
-        Assert.assertEquals("{\"value\":\"SECONDS\"}", text);
+        Assert.assertEquals("{\"value\":0}", text);
     }
 
     public static class Entity {
 
-        private TimeUnit value = TimeUnit.SECONDS;
+        private Integer value;
 
-        @JSONField(serialzeFeatures = { SerializerFeature.WriteEnumUsingToString })
-        public TimeUnit getValue() {
+        @JSONField(serialzeFeatures = { SerializerFeature.WriteMapNullValue, SerializerFeature.WriteNullNumberAsZero })
+        public Integer getValue() {
             return value;
         }
 
 
-    }
-    
-    public static enum TimeUnit {
-        SECONDS, MINUTES
     }
 }
