+++ src/test/java/com/alibaba/json/bvt/feature/FeaturesTest4.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt;
+package com.alibaba.json.bvt.feature;
 
 import org.junit.Assert;
 import junit.framework.TestCase;
@@ -8,14 +8,14 @@ import com.alibaba.fastjson.annotation.JSONField;
 import com.alibaba.fastjson.serializer.SerializeConfig;
 import com.alibaba.fastjson.serializer.SerializerFeature;
 
-public class FeaturesTest5 extends TestCase {
+public class FeaturesTest4 extends TestCase {
 
     public void test_0() throws Exception {
         SerializeConfig config = new SerializeConfig();
         config.setAsmEnable(false);
         
         String text = JSON.toJSONString(new Entity(), config);
-        Assert.assertEquals("{\"value\":false}", text);
+        Assert.assertEquals("{\"value\":\"\"}", text);
     }
     
     public void test_1() throws Exception {
@@ -23,15 +23,15 @@ public class FeaturesTest5 extends TestCase {
         config.setAsmEnable(true);
         
         String text = JSON.toJSONString(new Entity(), config);
-        Assert.assertEquals("{\"value\":false}", text);
+        Assert.assertEquals("{\"value\":\"\"}", text);
     }
 
     public static class Entity {
 
-        private Boolean value;
+        private String value;
 
-        @JSONField(serialzeFeatures = { SerializerFeature.WriteMapNullValue, SerializerFeature.WriteNullBooleanAsFalse })
-        public Boolean getValue() {
+        @JSONField(serialzeFeatures = { SerializerFeature.WriteMapNullValue, SerializerFeature.WriteNullStringAsEmpty })
+        public String getValue() {
             return value;
         }
