+++ src/test/java/com/alibaba/json/bvt/feature/FeaturesTest5.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt;
+package com.alibaba.json.bvt.feature;
 
 import org.junit.Assert;
 import junit.framework.TestCase;
@@ -8,14 +8,14 @@ import com.alibaba.fastjson.annotation.JSONField;
 import com.alibaba.fastjson.serializer.SerializeConfig;
 import com.alibaba.fastjson.serializer.SerializerFeature;
 
-public class FeaturesTest4 extends TestCase {
+public class FeaturesTest5 extends TestCase {
 
     public void test_0() throws Exception {
         SerializeConfig config = new SerializeConfig();
         config.setAsmEnable(false);
         
         String text = JSON.toJSONString(new Entity(), config);
-        Assert.assertEquals("{\"value\":\"\"}", text);
+        Assert.assertEquals("{\"value\":false}", text);
     }
     
     public void test_1() throws Exception {
@@ -23,15 +23,15 @@ public class FeaturesTest4 extends TestCase {
         config.setAsmEnable(true);
         
         String text = JSON.toJSONString(new Entity(), config);
-        Assert.assertEquals("{\"value\":\"\"}", text);
+        Assert.assertEquals("{\"value\":false}", text);
     }
 
     public static class Entity {
 
-        private String value;
+        private Boolean value;
 
-        @JSONField(serialzeFeatures = { SerializerFeature.WriteMapNullValue, SerializerFeature.WriteNullStringAsEmpty })
-        public String getValue() {
+        @JSONField(serialzeFeatures = { SerializerFeature.WriteMapNullValue, SerializerFeature.WriteNullBooleanAsFalse })
+        public Boolean getValue() {
             return value;
         }
