+++ src/test/java/com/alibaba/json/bvt/feature/FeaturesTest6.java
@@ -1,4 +1,6 @@
-package com.alibaba.json.bvt;
+package com.alibaba.json.bvt.feature;
+
+import java.util.List;
 
 import org.junit.Assert;
 import junit.framework.TestCase;
@@ -8,14 +10,14 @@ import com.alibaba.fastjson.annotation.JSONField;
 import com.alibaba.fastjson.serializer.SerializeConfig;
 import com.alibaba.fastjson.serializer.SerializerFeature;
 
-public class FeaturesTest3 extends TestCase {
+public class FeaturesTest6 extends TestCase {
 
     public void test_0() throws Exception {
         SerializeConfig config = new SerializeConfig();
         config.setAsmEnable(false);
         
         String text = JSON.toJSONString(new Entity(), config);
-        Assert.assertEquals("{\"value\":0}", text);
+        Assert.assertEquals("{\"value\":[]}", text);
     }
     
     public void test_1() throws Exception {
@@ -23,15 +25,15 @@ public class FeaturesTest3 extends TestCase {
         config.setAsmEnable(true);
         
         String text = JSON.toJSONString(new Entity(), config);
-        Assert.assertEquals("{\"value\":0}", text);
+        Assert.assertEquals("{\"value\":[]}", text);
     }
 
     public static class Entity {
 
-        private Integer value;
+        private List value;
 
-        @JSONField(serialzeFeatures = { SerializerFeature.WriteMapNullValue, SerializerFeature.WriteNullNumberAsZero })
-        public Integer getValue() {
+        @JSONField(serialzeFeatures = { SerializerFeature.WriteMapNullValue, SerializerFeature.WriteNullListAsEmpty })
+        public List getValue() {
             return value;
         }
